FastAPI-backend/requirements.txt
fastapi
uvicorn
web3
py-solc-x
streamlit
python-dotenv
requests
pandas
pydantic
python-multipart
scikit-learn
langchain-groq
langchain-core